move login command to yargs
